fix: optimised graceful shutdown to support `sync`
